pyproject.toml
[project]
name = "my-flask-templates"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask-sqlalchemy>=3.1.1",
    "flask>=3.1.1",
    "python-dotenv>=1.1.0",
    "flask-migrate>=4.1.0",
    "flask-cors>=6.0.0",
    "flask-mail>=0.10.0",
    "gunicorn>=23.0.0",
    "flask-jwt-extended>=4.7.1",
    "captcha>=0.7.1",
    "flask-redis>=0.4.0",
    "resend>=2.10.0",
    "pymysql>=1.1.1",
    "flask-wtf>=1.2.2",
    "flask-restful>=0.3.10",
    "flask-openapi3[elements,rapidoc,rapipdf,redoc,scalar,swagger]>=4.1.0",
    "setuptools>=80.9.0",
]
